hack for other compatible fp sensors
dirty hack for compatible fingerprint sensors with capacity greater than the default, since the function limits reading the templates to 162(?).  an observation: each pass reads 256 templates (0-255) so for sensors with 1000 templates, 4 passes are required.  but there must be a better way to do this.  tested on this hardware: https://circuit-help.com.ph/product/fingerprint-sensor-2/